requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL
dj-database-url==2.1.0

# WebSocket for Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# Blockchain (USDT Integration)
web3==6.11.3
eth-account==0.10.0

# API Documentation
drf-yasg==1.21.7  # Swagger UI
drf-spectacular==0.27.0

# Utilities
python-decouple==3.8  # Environment variables
Pillow==10.1.0
python-dotenv==1.0.0

# Payment Processing
stripe==7.8.0  # Optional: for fiat payments

# Monitoring & Logging
sentry-sdk==1.39.1

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.18.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0  # Static files

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Date/Time
pytz==2023.3
python-dateutil==2.8.2
